debian.master: re-enable AUFS filesystem in kernel
AUFS became disabled in Ubuntu 22.04, but SPS uses it for merging a
writable layer onto the firmware on squashfs, and we would rather not
switch to overlayfs at this point yet.

With this change the AUFS feature can be turned on with 'aufs' kernel
module as before.

References: jira PAM-16796