debian.master/abi/amd64/ignore: Skip ABI check
Contrary to the comment in debian/rules.d/4-checks.mk, checking
the ABI for changes is relevant exactly when the ABI number
(between the hyphen and the dot in the full version) does NOT
change. Actually, debian/scripts/abi-check enters "no-fail" mode
when the ABI numbers being compared are different, which is the
correct behavior.

Unfortunately official Ubuntu kernel source releases often
contain outdated CRC info, which the ABI checker uses as
baseline anyway, ignoring the origin version noted in the
debian.master/abi/abiname file: it just assumes the ABI
numbers passed in by the makefile. The makefile derives
these from the last two entries in debian.master/changelog,
so this causes no problems most of the time, because these
are usually different in official Ubuntu releases, so the
"no-fail" mode is activated and the build does not break
even with outdated CRC info.

For example at tag debian/5.15.0-92.102 the abi-check finds
26325 symbols in the shipped debian.master/abi/amd64/generic
(which corresponds to ABI number 91) and the same number
from debian.master/__abi.current/amd64/generic (which is
generated during the build, so ABI number 92), and it reports

  WW: 2597 symbols changed hash and weren't ignored

but it still exits with status 0, so the build concludes
successfully. However, just but adding a new entry to
debian.master/changelog with version 5.15.0-92.102bb1 (like
we routinely do after each merge) results in

  $ debian/rules abi-check-generic
  [...]
  EE: 2597 symbols changed hash and weren't ignored
  [...]
  make: *** [debian/rules.d/4-checks.mk:4: abi-check-generic] Error 1

In other words, the same differences are considered errors
now because of the identical ABI numbers in the last two
changelog entries, and the build fails.

The same problem does not arise at tag debian/5.15.0-91.101
for example, but this would be too hard to track on a case
by case basis, and we don't compile out-of-tree modules to
the SPS kernel anyway, so ABI stability has little value
for us, so we decided to ignore ABI changes in our Linux
kernel builds.

Aside, our current delta against Ubuntu (enabling AUFS and
FIPS) does not break the kernel ABI, just extends it with
three symbols.

References: azure #444668
References: https://wiki.ubuntu.com/KernelTeam/KernelMaintenance#Overriding_ABI_check_failures
References: https://wiki.ubuntu.com/KernelTeam/AbstractedDebian